requirements.txt
opencv-python
screeninfo
asyncio
requests
python-dotenv